refactor: delete folder
