+++ src/app.module.ts
@@ -5,7 +5,7 @@ import { UsersModule } from './users/users.module';
 import { TracksModule } from './tracks/tracks.module';
 import { ArtistsModule } from './artists/artists.module';
 import { AlbumsModule } from './albums/albums.module';
-import { FavoritesModule } from './favorites/favorites/favorites.module';
+import { FavoritesModule } from './favorites/favorites.module';
 
 @Module({
   imports: [
