+++ src/favorites/dto/create-favorite.dto.ts
