+++ src/favorites/favorites.model.ts
