+++ src/favorites/interfaces/favorite.interface.ts
